pyproject.toml
[tool.poetry]
name = "fast-api-todo"
version = "0.1.0"
description = ""
authors = ["Davison Silva <davison.vinicius360@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.114.0"}
sqlalchemy = "^2.0.34"
alembic = "^1.13.2"
pydantic-settings = "^2.4.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.6.4"
pytest = "^8.3.2"
pytest-cov = "^5.0.0"
taskipy = "^1.13.0"
httpx = "^0.27.2"

[tool.ruff]
line-length = 79
extend-exclude = ['migrations']

[tool.ruff.lint]
preview = true
select = ['I', 'F', 'E', 'W', 'PL', 'PT']

[tool.ruff.format]
preview = true
quote-style = 'single'

[tool.pytest.ini_options]
pythonpath = "."
addopts = '-p no:warnings'

[tool.taskipy.tasks]
lint = 'ruff check .; ruff check . --diff'
format = 'ruff check . --fix; ruff format .'
run = 'fastapi dev fast_api_todo/app.py'
pre_test = 'task lint'
test = 'pytest -s -x --cov=fast_api_todo -vv'
post_test = 'coverage html'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
